Show share link when clicking on icon
Clicking on the share icon didn't result in the share link being shown.
For now, the simplest way to fix this is to ensure that we get the
correct container element by using event.currentTarget (which will
always be the <a> tag) rather than event.target (which might be the <i>
tag).