Move Hypothesis API access into api_client module
Add an api_client.get() function that wraps making a GET request to the API.

Also add a config.api_url() function that wraps getting the h.api_url setting
and falling back to a default.

Moving this stuff out of view code makes it reusable, and the view code
simpler.