AnnotionUISync: export show/hide frame
Currently, the show/hide frame functionality is not directly
available for the sidebar code.

However, internally, then annotation-ui-sync service implements
the necessary features.

This commit exposes those implementations as methods on the AnnotationUI
service, which is the top-level entry point for all similar APIs for
controlling the state of the UI.