Don't use webassets cache in development
The utility of the cache has gone down since we started using
browserify because the results of intermediate coffee compilations
aren't handled by webassets anymore. It is still the case that
webassets will compare timestamps, so things will not rebuild unless
changes are made, as before. The only difference now is that CSS
builds might be slightly slower after changes, but I think that's
actually better since I've seen issues with the cache not being
intelligent enough.