Refactor h.stream.atom.limit setting handling
We validate the setting from the config file at boot, convert it to an
int and insert a fallback value into the settings if one isn't there.

Crash at boot if the value is invalid.

Request-time code assumes there will be a setting in the settings to
fall back to if there's none in the params. Crashes if there's an
invalid value in the params.