Move CSRF handling into a tween
This removes the need for various components to worry about whether
they have or need a fresh csrf token. With the separation of the
site and the api into different apps it's no longer a useful
useful optimization (as if it ever was, really) to try to manage
this for only some routes.