Automatically scroll to new annotations
When the user adds a new annotation scroll the sidebar to show it.

The way this has been implemented causes annotations to be scrolled to
when their edit buttons are clicked as well, not just when new
annotations are created, which causes things to jump around confusingly
when clicking edit buttons. This needs to be fixed.

This fixes #2053 - in cases when there are more than a screenful of
annotations in the sidebar, creating a new annotation causes a new
annotation to appear in the sidebar but this may happen off screen
leaving the user confused. (For example: scroll to the very bottom of
the sidebar then create a new annotation at the top of the page, or
vice-versa.) The fix is to scroll the sidebar to show the new
annotation.

The diff for annotation.html is a mess - we've only actually changed a
couple of things, but we had to change the indentation of the while file
so git thinks we changed everything.

- We use #fragment links to scroll to the new annotations.
- This requires the HTML elements for the annotations to have unique id
  attributes.
- We can't use the IDs of the annotations from the server because new
  annotations that haven't been saved don't have IDs yet.
- So instead we use timestamps.
- We use a new scrolltome directive that scrolls the page to its element
  when instantiated.
- We attach this scrolltome directive to an element within the
  annotation that has ng-if="vm.editing", whenever this expression
  becomes true the directive will be instantiated and cause a scroll.
  The expression is true when new annotations are added.
- We wrap the HTML for an annotation in a new div to hold the id that
  we're going to link to. We need an element that contains all of the
  HTML for the annotation so that the whole annotation is always
  scrolled into view, whether we're scrolling up or down to get to it.
  We can't use the outermost <article> of the annotation in thread.html
  because that's not within the scope of the AnnotationController where
  the element_created_at property that we need is located. So we add a
  new element.