Add Google Analytics to single-page application
A `$rootScope.$apply` is removed from the `on` method of the iframe-
side cross-frame code. This code was incorrect because the registration
of a handler does not require a digest cycle. This was seemingly
innocuous before, but adding the analytics module changed the order
of injection resolution such that it caused a problem wherein a digest
cycle happens before angular has even finished compiling the DOM and
processing ng-template scripts. The result was that route templates
were not being read from the template cache and 404s resulted on
stream navigation and auth pages.