Be strict about up/down arrow offscreen counts
It is potentially confusing if, upon reaching the start or end of
highlights, the navigational buckets don't seem to do anything. At this
point, the first or last highlights are visible but the buckets seem to
indicate that the highlights are still off screen.

Change the buckets to mantain a count of the annotations that are
strictly off screen.