Improve make targets
dev depends on deps now and deps depends separately on h.egg-info
and node_modules so that there can be something actually modified
on the filesystem for make to compare times. These build outputs
are touched by their rules and depend on package.json and setup.py
so they should be regenerated whenever our deps change.