Send a redirect for app.html -> viewer
In particular, this seems to fix an issue with Angular in Firefox
that causes an infinite digest cycle due to the router's redirect
rule. It's a bit mysterious and there are several issues that have
been opened and closed on the Angular issue tracker, but no solution.

Perhaps a newer version of Angular.