Replace gnsq with nsq-py
nsq-py seems at least as well maintained, it is closer in API to
the official pynsq client, supports TCP connections to nsqd,
and supports the iteration protocol as a way to consume Reader
messages, which is simpler than the signal-based callbacks and
feels a bit more Pythonic.

Finally, it doesn't depend on gevent, and it's always nice to bake
that assumption into fewer parts of the system.