Don't return JSON with HTTP redirects
This reverts a change made in 39d1468 which assumes that a view returns
JSON along with an HTTPRedirection response. This breaks logout (and
other views) that return empty redirects.

In addition, because the edit_profile view function no longer returns a
redirect, the AsyncFormViewMapper no longer needs to preserve JSON
responses in this fashion.