Dehorusify the ForgotPasswordController
Removes the ForgotPasswordController's dependency on horus. This makes
it easier to see what is going on when we reset passwords, as well as
much easier to test the controller view function. As such, this commit
also adds some tests for both the forgot_password and reset_password
view functions.

For the time being this will break the rendered forms at
/forgot_password and /render_password. (But the ajax forms from the
sidebar will continue working.) I will bring back this functionality in
a subsequent commit.