Remove support for horus config flags from h.accounts.schemas
This removes the ability to toggle the requirement for user activation,
and in so doing simplifies validation of the login form.

While the ability to disable the activation requirement in development
is convenient, after discussion with Sean we agreed that if we're doing
it in prod we should be doing it in development. Seeing as activation
emails are printed conveniently to STDOUT when running in development
mode, this seems reasonable.