Remove unnecessary workaround for horus bug
This code worked around a bug in horus where the activation email would
be generated before the new user was flushed to the database, resulting
in a broken activation link.

Seeing as we now control the code that generates said email, and we are
currently flushing at an appropriate point (as it were...) we no longer
need this.