Deduplicate server-side password validation
The same code for creating a colander.SchemaNode for a new password was
duplicated four times in accounts/schemas.py and with some differences
between the copies.

This meant that for example you could not register a new account with a
single-character password, but you could use the edit profile form to
change your password to a single-character after registering.