Remove "at least two characters" password form hints
These duplicate the fact that a password needs at least two characters
(if we wanted to change that, we would need to find and change the
number two in multiple places in the UI) and they just tempt people to
choose "aa" for their password. If people are going to try to submit
single-character passwords I'm happy for them to have to submit the form
to find out that it's not allowed.