Add NIPSA (Not In Public Site Areas) feature
This feature allows admin users to flag users as "nipsa'd", meaning that
these users' annotations will not appear in "public site areas".
Currently "public site areas" are defined as basically everywhere:
nipsa'd annotations are filtered out from the results returned by the
search API, so they won't appear in the stream or in the sidebar for
example.

- h/api/nipsa/ contains a new nipsa package with:
* nipsa table in db that records ids of nipsa'd users
* search functions for inserting nipsa-filtering commands into Es query
  dicts
* logic functions for using the nipsa feature, e.g. for nipsa'ing or
  unnipsa'ing a user. or asking whether a user is nipsa'd
* a worker function that adds the nipsa flag to or removes the nipsa
  flag from all of a user's past annotations

h/api/search.py: Now runs Es queries through the nipsa-filtering function

h/api/views.py: When creating a new annotation, if the user is nipsa'd
it now adds the nipsa flag to the annotation.

h/nipsa.py: A simple HTML page for admin users to nipsa and unnipsa
users.

Note on app/API separation: We don't want the app to read or write
database tables belonging to the API or vice-versa, nor do we want the
API to import modules defined in the app. The obvious place to put the
nipsa flag would be in a column on the user table (which belongs to the
app), but then when the user creates a new annotation the app would have
to communicate somehow to the API that the authorized user is nipsa'd
without the API reading from the user table or calling a function
defined in the app. Therefore the nipsa flag has been stored in a
separate nipsa table owned by the API instead.