Add h.db module, providing db session and base class
We currently use pyramid_basemodel to set this up for us, but it's quite
a lot of code, and we're not using more than a tiny fraction of it. More
importantly, the mechanism by which it configures database sessions
doesn't allow us to have multiple distinct sessions within one
application (i.e. for the app and API).

This commit introduces the small amount of code necessary to manage our
own database session and base class, using sqlalchemy's `scoped_session`
helper.

The current session is always available as `h.db.Session`, but access
like this is unlikely to be necessary, as model classes should be able
to use their own `query` property, which is dynamically bound to the
session. Direct access to the session from view functions will be done
via the request property `db`, also provided by this module.