Use ng-bind rather than templates to avoid rendering errors
Use of {{angular templates}} causes problems on pages with Annotator
enabled due to Annotator's habit of splitting text nodes whenever you
select anything. Specifically -- if you select one of the words in this
text and then change the privacy setting, it will end up rendering as
something like

   PubliOnly Me
   Only Melic

or similar. Using a span tag and ng-bind is a bit of a hack but it
prevents this from happening.

Oh, Angular...