Make feature flags runtime toggles stored in the database
The way we had implemented feature flags was pretty inflexible, and
wasn't effectively allowing us to do the one thing that feature flags
are really intended to do, namely:

> Decouple deployment (when you push code to production) from release
> (when users see new features).

A truly effective feature flag system is one which achieves the above,
and also enables more dynamic options for "releasing" such as:

- release to admin users/beta users only
- release to a randomly selected 10% of users
- release only to one specific user

This requires that the feature flag system have some knowledge of the
accounts system.

Furthermore, in order to make the feature flag system truly atomic and
reversible (i.e. shit has gone wrong, let's flip that back) we need to
be able to toggle feature flags instantly, without needing to wait for
an application redeploy.

Both of these requirements indicate a feature flag system fully
integrated with the request-response cycle of the application, and one
where flag status can be more complex than a simple boolean.

In order to make this possible, this commit puts feature flags into a
"feature" table of the main application database. This allows us to
extend the definition of flag state as and when we see fit to enable
some or all of the release options denoted above.

The initial state of all feature flags is off, but in order to
facilitate an easy migration path for us the data migration with this
commit will read the current state of the existing flags from the system
environment and write appropriate rows into the table.