Add a minimal feature client to the front-end
In order to respond to feature flag changes on the frontend, we need a
way of knowing what the current feature flag values are on the backend.

This commit adds a basic feature flag client that retrieves current
feature flag values by making an ajax request to the server. The client
also includes a cache, and supports retrieving new value from the server
when the cache expires.