Add URI normalisation behind a feature flag
This is the first step in a major cleanup of how we internally normalise
URIs referenced by annotations. The URIs in question divide into two
groups:

1. URIs stored on annotations which refer to the annotated resource.

2. URIs discovered in the annotated resource or elsewhere, which can be
   used as hints to tell us which other URIs refer to the same
   underlying document (e.g. to alternate formats of the annotated
   resource).

We use both sets of these URIs when a user visits a page and activates
Hypothesis:

- the client sends a search request, which includes the URI of the page
  on which it was activated.
- the server must find all annotations made on that page, but *also* all
  annotations made on alternate versions of the page, and return these
  to the client.

Getting this right is both important and hard.

It's important because poor implementation will result in a poor
experience of using Hypothesis. Users don't (usually) want to see
annotations from documents other than the one they're viewing, and users
(usually) don't want any of the annotations on the page they're viewing
to be missing.

It's hard because in simple terms, a user expects that when they visit
the "same page," they will see the same annotations. But "same page" is
not a well-defined concept on the web.

Indeed, one of the so-called "axioms of web architecture" is the
[opacity of URIs](http://www.w3.org/DesignIssues/Axioms.html#opaque),
which means that a priori, there are only a limited number of
normalisations of a URI which are guaranteed to preserve reference.

At the time of writing, the standard which defines permitted
normalisations is RFC 3987[1], AKA the "Internationalized Resource
Identifiers (IRIs)" RFC.

[1]: https://tools.ietf.org/html/rfc3987#section-5

Dealing correctly with internationalised identifiers is something we
will probably need to do in due course. For now, dealing correctly with
URIs will be a substantial improvement on the status quo. IRIs won't be
referred to again. The relevant standard which defines permitted
normalisations of URIs is RFC3986[2].

[2]: https://tools.ietf.org/html/rfc3986#section-6

It's worth reading the above if you're trying to understand the starting
point for this code, but it's also worth understanding that in reality,
we have to be a lot more "heuristic" than these documents allow if we
want to ensure a good user experience. The "opacity of URIs", sadly,
must go out the window (for URLs, at least) if we want to provide a good
experience with Hypothesis on the real web.

For example, in practice, it may be the case that lexically sorting
URL query parameters is almost always a permissible (and thus
experience-improving) normalisation. But such an operation is clearly
not permitted by the URI normalisation specifications.

So, overall, here's the approach taken by this commit:

- We introduce the concept of a URI normalisation, which maps an input
  URI idempotently to a normalised version of that URI.

- When storing new annotations, we store normalised versions of the
  'source' property of targets (that identify the annotated document).

- When searching via the API:

  + the value associated with the "uri" keyword, if passed, is used to
    look up alternate URIs.[^1]

  + Together, the query URI and alternate URIs returned from the
    previous step form a set for searching.

  + Each of the URIs in the search set are normalised.

  + The annotation database is queried to find those annotations where
    the normalised target source field is an exact match for any of the
    resulting set of normalised URIs.

- We continue to store, unmodified, the URIs as passed by the client, so
  that we can, if necessary, alter the URI normalisation operation in
  the future.

A later commit will introduce a mechanism to regenerate the normalised
fields in the search index to facilitate both release of this feature in
a deployment with legacy annotations lacking such fields, but also the
ability to later update the URI normalisation procedure.

The addition of normalised fields to annotations is performed
unconditionally, but the alternate search procedure is guarded by a
feature flag.

[^1]: Ideally we would look up such "URI equivalences" using a
normalised URI, but this is complicated by the implementation of
`annotator-store` at the moment, so we don't yet do so.