Straightforward API inclusion
Clean up the boundaries between h.api and the rest of the application
so that it can be included in a straightforward manner.

The h.api module becomes more or less oblivious to authentication and
authorization, except that it still defines ACL properties for its
resources and it defines the token route.

The token route can be ripped out next and placed in another module,
since the h.api module needn't even know that OAuth tokens are in use
and in this way it can cease to rely on `request.create_token_response`.

This commit reintroduces pyramid_multiauth as a straightforward way to
support SessionAuthenticationPolicy and RemoteAuthenticationPolicy in
the same application.

The case of the default client in h.auth and the session view callable
are now oblivious to type of authentication policy in use.