Add a group package and data model
This commit adds a package "h.groups", which for now contains only data
models for group membership, and a schema migration to create the
relevant tables in the database.

For now, groups are identified simply by unique id, and they have a
unicode name, which is expected to be their display name.