Enable nipsa'ing users who don't exist
I don't think it's a good idea to restrict nipsa to users who we have a
row for in the users table. The model (nipsa'd user IDs in a separate
table with no relation to the user table) allows us to be more flexible
and nipsa any userid. Taking advantage of that avoids potential future
problems such as:

- Nipsa'ing users whose accounts have been deleted (if in the future we
  enable deleting accounts and it removes the row from the users table)
- Nipsa'ing users from third-party account providers (if in the future
  we enable external accounts and it's implemented in such a way that
  the users have no row in our users table)

So nipsa is just a simple list of userids. Each id may or may not exist
in our users table as well, it doesn't matter. And any annotations
created by these ids are hidden.