Filter new nipsa'd annotations from the WebSocket
When a nipsa'd user creates or updates an annotation this creates an NSQ
event that goes through streamer.py and is sent via a WebSocket to any
users viewing the same page. Those users will see the annotation appear
in their browser without refreshing the page.

Don't send nipsa'd annotations down the WebSocket in this way, except to
clients where the logged-in user is the nipsa'd user who created the
annotation (nipsa'd users can see their own annotations).