Make Makefile more dumb
Having a clever Makefile that tries to work out when it needs to install
updated dependencies is a minor convenience for those of us working with
`h` on an ongoing basis.

But sometimes that process goes wrong. If for whatever reason the
`h.egg-info` file ends up more recent than `setup.py`, users get
confused by the Makefile not doing what it's supposed to.

This commit removes all magic:

- `make deps` installs deps, unconditionally
- `make dev` runs gunicorn

Our [hacking guide][1] already instructs users to run these two
separately.

[1]: https://h.readthedocs.org/en/latest/hacking/install.html

Closes #2391.