Hook up the new group form to the model
On form submission create a new group in the db and then redirect to the
group's page.

Validation of form params still needs to be done.

We're using the group id direct from the db in the group page's URL,
this needs to be replaced with a hashid based on the group's id and some
salt.