Consolidate and style the admin interface
This commit brings together "administrator" views into one place -- the
Hypothesis admin console, found at /admin.

So far, the only to admin views we have are for flagging/unflagging
users as NIPSA, and for administering administrative user status. In the
near future, we may also have a feature flag console in here.

- I've moved all the admin views into `h.admin` so that we don't
  confused business logic view functions with administrative view
  functions, and so that in future it is easier to treat the `/admin`
  tree as a single entity for the purposes of permissions, etc.
- I've added an "admin" page layout, which uses Twitter Bootstrap for
  basic styling.
- I've rejigged both the NIPSA and the administrator users pages to make
  use of bootstrap.