Deduplicate app bootstrap in h.script
Almost every one of these commands needs to bootstrap the application
and get access to the request object before proceeding. This commit
pulls that into the `bootstrap` function which each command can then
use.

N.B. That the order of arguments to `hypothesis admin` has changed here
(from 'username config_uri' to 'config_uri username') in order to make
it consistent with all the other commands.