Add `hypothesis annotool` for modifying the annotation database
This commit adds a command to the `hypothesis` CLI, which enables
running of functions which modify annotations, so that we can run data
migrations against the annotation database.

The only supported function at the moment is 'normalise-uris', which
ensures that all annotations in the database have normalised copies of
their "target.source" attributes.