Fix a pep257 warning
h/db.py WARNING: __all__ is defined as a list, this means pep257 cannot
reliably detect contents o f the __all__ variable, because it can be
mutated. Change __all__ to be an (immutable) tuple, to remove this
warning. Note, pep257 u ses __all__ to detect which definitions are
public, to warn if public definitions are missing docstrings. If __all__
is a (mutable) list, pep257 cannot reliably assume its contents. pep257
will proceed assuming __all__ is not mutated.