URI normalisation: normalise percent-encoding
Normalise percent-encoding of paths and query strings as dictated by
RFC3986.

Paths and query strings can both contain [percent-encoded][1] character
data. Percent-encoding is a way to include  in URLs (and URIs more
generally) certain characters which may have semantic meaning in a URL
(such as "/" in a URL path), or which may simply not be permitted (such
as literal UTF-8 characters).

Sometimes these percent-encodings must be preserved -- percent-encoded
UTF-8 unicode, for example. But sometimes URLs that differ only in
percent-encoding can (according to RFC3986) be considered the same. To
pick one example, "-" (hyphen) is an "unreserved character", so:

    http://example.com/foo-bar, AND
    http://example.com/foo%2Dbar

can be considered the same. In addition, case in percent encodings does
not carry information, so

    http://example.com/foo%2dbar, AND
    http://example.com/foo%2Dbar

are also the same URL.

This commit adds support for normalising percent-encoding of URL paths
and query strings.

[1]: https://en.wikipedia.org/wiki/Percent-encoding