Only use PDF.js if annotating
PDF.js Viewer + embed.js loads for browserAction toggle.
However, it uses the old URL sniffing approach when a PDF
is currently displayed on the tab when the browserAction
button is clicked--which means, in that specific scenario
only URLs containing `pdf` will get reloaded into the
PDF.js Viewer.

That said, this does handle all other scenarios, such as:
 - currently annotating and pasting in a URL to a PDF
 - browsing to a PDF while annotating
 - toggling annotating on while viewing a URL containing `.pdf`