Support for server sent events for annotations
The implementation relies on Elasticsearch percolator documents and
ephemeral nsq topics. Each client creates an ephemeral topic and a
percolator. The percolator stores the query from the request parameters
and whenever writes occur the annotation document is percolated and
sent to all matching topics.

The TTL feature of Elasticsearch ensures that percolator documents
are cleaned up even in the event of a server fault.