Use a simple_query_string for search remainders
Specifying multiple "any" terms in a search was resulting in only the
last term being applied, since "multi_match" takes a string. Instead,
change to the "simple_query_string" query and join the terms. This
allows multiple terms to be specified and enables the user to use
some advanced, but fairly safe, constructions.