Flexbox to the rescue
Lay out the search bar and the top bar using flexbox because the
vertical centering is much better and easier using flexbox instead
of floats and line heights.

As long as we're using only features easily supported in the 2012
and final spec the -ms prefix gets us IE10. All other browsers we
care about should be supported already by the compass mixins.