Get username from auth not from form
When editing a user's profile or disabling a user, use the user from
request.authenticated_userid not the username submitted in the form.

Remove username from the schema so the form submission doesn't need to
contain it.

The change email form was getting a validation error back from the
server because it didn't submit the username in the form - now it
doesn't need to. Fixes #2441.