Make api.views.search() a wrapper
Move the code into h.api.logic.search_annotations() and make
h.api.views.search() just a view-callable wrapper for this.

This is so that logic.search_annotations() can be a function that other
packages within our codebase can import and call, and views.search() is
just a wrapper to make search_annotations() available via the HTTP/JSON
API as well.