Let the compass filter do the url rewrites
This is a little bit confusing, but it is better this way. It's
confusing because webassets doesn't set the compass image directory,
so the paths still have an "images/" prefix. However, the cssrewrite
filter uses the output URL from webassets and that will be an absolute
URL at request time. The compass filter uses only the base URL from the
environment configuration, which is relative. Letting compass do the
rewriting therefore results in merged compass files having correct, but
relative, paths regardless of whether the assets are generated at the
command line or at request time. This should make it possible to pre-
compile the assets before the first request without generating bad
URLs, and to virtual host the application under multiple URLs.