Use permissions to decide who can read annotations
Use an annotation's 'permissions' field (instead of its 'group' field)
to decide whether a request can read the annotation.

This is for the API read() view / the annotation's individual page, not
for showing the annotation in search results.

This seems to break search - group annotations aren't showing up in
search results even for users who are group members.

On annotation create or update we set the annotation's read permissions
according to its group. Basically if it's a shared (not user-private)
annotation within a group then its ['permissions']['read'] is set to
['group:<hashid>']. Private and non-group annotations don't change.

On annotation read the view function calls an auth function that checks
whether the request has an effective_principal that's in the
annotation's read permissions. For group annotations this means you have
to be a member of the group to read them.