Move update annotation auth check out of view
Move the authorization check when updating or deleting an annotation out
of the view callables and into the Annotation resource object. Pyramid
will not even call the view callable, if the auth check fails.

When creating or updating an annotation, if the annotation belongs to a
group then set annotation['permissions']['update'] to
'<userid>~group:<hashid>'.

Add '<userid>~group:<hashid>' principals to request.effective_principals
for each group that the user is a member of, and 'group:__none__'.

Hack permissions.coffee to show the edit button if the annotation has an
ACE whose principals starts with the user's id.

Fix a thread-safety issue in transform.py.

Delete h.api.groups.auth.authorized_to_write_group(), no longer needed.