Unbundle URL polyfill
The js-polyfills library is a nice one. It provides enough for
PhantomJS and it provides the URL polyfill. However, we keep core-js
since for production it's better to have the lightweight requirement
of just the es6.promise module without global namespace pollution.

We tolerate pollution of the global namespace for browsers with a
broken or absent URL constructor implementation.