Refactor Pyramid routing and view configuration
When resources are backed by database models, keep the two separate
to improve extensibility and package isolation. By separating the
resource tree from the persistence models, the resource tree can be
extended without dealing dealing with database models and mappers.
With sub-classing of the resource tree, explicit containment of the API
is no longer needed for any view other than the exception views. The
annotation model also becomes simpler, no longer containing Pyramid
permissions details. In this manner, more details about groups and
users are lifted out of the annotation API package.

Remove the "admin party" mode for annotations since otherwise this
would introduce "upsert" operations using PUT for anyone in any group.
Simply removing this unused functionality allows the annotation resource
to be behave as if its model always exists, even if empty and not
persisted, which avoids having to deal with that case within the acl code
and the view callables.

Factor out a generic Collection resource and implement the stream and the
search in terms of that. Remove most of the logic from the stream view
by making the tag and user search shortcuts implemented as a collection of
collections of annotations so that the view itself need only inspect its
context for the presence of a query. Future work can remove the need for
a redirect by teaching the client about these routes, thereby beautifying
the stream URLs for real user and tag pages.

Make all the view configuration decorators and helpers more concise and
consistent.