Remove the 'group' field from annotations
Remove the 'group' field and relax the server munging. Derive
everything from the ACL for now and have create/update validate
that, for now, every principal named in the 'read' permission is
effective for the current request.

In other words, the client can provide whatever values it wants for
admin/update/delete, but it can only expose an annotation to groups
the user belongs to.