Copy parent scopes into replies
Copy annotation.target.scope from parent annotations into reply
annotations in prepare(). This means that replies show up in searches
for annotations of a given URI (searches for annotations with a given
URI in their scopes), such as when viewing the annotations of a page in
the sidebar, when the search_normalized feature is active.