Make it easier to use hashids
This commit gives the hashids module a global salt set at application
boot-time, which makes it substantially simpler to compute a hashid for
a group without access to the request object.

There is obviously a problem with this, which is that the h.hashids
module is no longer usable in multiple different applications at the
same time, with different salts. But I think we'll live with that.