Relax restrictions for annotations with no group
If the stored annotation doesn't have a group, don't worry about
the group field. We will want to run a migration, perhaps, to give
each annotation a default group eventually when we roll this out.
In the meantime, it's problematic to have the render function fill
in a default group and then have this enforcement here.